implement basic cli streaming inference. add basic apply_template func.
